BE-723 | Fix 0 liquidity pools for candidate routes
Some candidate routes have their capitalization set to 0 because that's how it's defined in the block data. As a result, the router avoids routing through these pools.
This PR fixes the issue by assigning a liquidity capitalization value to candidate routes with 0 capitalization. The value is retrieved from the pools use case, which provides the most up-to-date data.